se realiza el crud en los controladores

+++ src/main/java/com/hotel/repositorio/ReservaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hotel.modelo.Reserva;
 
-public interface ReservaRepositorio extends CrudRepository<Reserva, Integer> {
+public interface ReservaRepositorio extends CrudRepository<Reserva, Long> {
 
 }
