+++ src/main/java/com/hotel/repositorio/SalidaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hotel.modelo.Salida;
 
-public interface SalidaRepositorio extends CrudRepository<Salida, Integer>{
+public interface SalidaRepositorio extends CrudRepository<Salida, Long>{
 
 }
